refactor: GameRunner -> Racing, Cars -> RacingCars, OutputDTO -> RoundResultDTO로 명칭 변경
객체의 의도와 책임을 명확하게 하기 위해서 Cars -> RacingCars로 명칭 변경(Cars를 Car의 List에 대한 일급 컬렉션으로만 다룩기엔 책임이 너무 많음)
