+++ src/main/java/racingcar/dto/input/InputDTO.java
@@ -1,4 +1,4 @@
-package racingcar.dto;
+package racingcar.dto.input;
 
 public record InputDTO(String [] names, int trialCount) {
 }
