+++ src/main/java/racingcar/view/OutputView.java
@@ -1,19 +1,19 @@
 package racingcar.view;
 
 import java.util.List;
-import racingcar.dto.OutputDTO;
+import racingcar.dto.output.RoundResultDTO;
 
 public class OutputView {
     private static boolean isFirstTime = true;
 
-    public static void printRacingStatus(List<OutputDTO> outputDTOs) {
+    public static void printRoundResult(List<RoundResultDTO> roundResultDTOS) {
         if (isFirstTime) {
             System.out.println("\n실행 결과");
             isFirstTime = false;
         }
 
-        for (OutputDTO outputDTO : outputDTOs) {
-            System.out.println(outputDTO);
+        for (RoundResultDTO roundResultDTO : roundResultDTOS) {
+            System.out.println(roundResultDTO);
         }
         System.out.println();
     }
